Move ScriptPubKeyBufExt to addresses
The new_p2pkh function is necessary to implement the script_pubkey
function on Address. Since the function signature includes a hash type
from crypto, it can't be moved to primitives yet. Instead, we can move
the extension trait to addresses where it can be used, and re-export
the trait in bitcoin to retain the same API.

Move ScriptPubKeyBufExt to addresses, Adjust new_p2pk function impl to
remove use of Builder::push_key.
